fix(react-email): errors on renderAsync not being caught
